requirements.txt
Django==1.7
djangorestframework==3.5.3
MySQL-python==1.2.5
Pygments==2.1.3
South==1.0.2
